+++ CHANGELOG.md
@@ -6,11 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [3.0.0-beta1] - 2025-01-26
+- **Breaking change** Add `Ascii` and `Asciis` type.
+### Added
+- **Breaking change** Change return type of `Ascii()`.
+### Removed
+- **Breaking change** Remove `AsciiChars()` and `AsciiSpan()`.
+
 ## [2.5.1] - 2025-01-26
 ### Changed
 - Fix `ConsoleReader.String`
 ### Removed
-- Remove .NET 9 build
+- - Remove .NET 9 build
 
 
 ## [2.5.0] - 2025-01-26